@@ -1,5 +1,5 @@
 import React from 'react'
-import { AppBar, Toolbar, Typography, Button } from '@material-ui/core'
+import { AppBar, Toolbar, Typography, Button } from '@mui/material'
 import { Link } from 'react-router-dom';
 
 export default function AppBarred({ isLoggedin }) {
