@@ -1,18 +1,18 @@
 import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import Button from '@material-ui/core/Button';
-import Dialog from '@material-ui/core/Dialog';
-import ListItemText from '@material-ui/core/ListItemText';
-import ListItem from '@material-ui/core/ListItem';
-import List from '@material-ui/core/List';
-import Divider from '@material-ui/core/Divider';
-import AppBar from '@material-ui/core/AppBar';
-import Toolbar from '@material-ui/core/Toolbar';
-import IconButton from '@material-ui/core/IconButton';
-import Typography from '@material-ui/core/Typography';
-import CloseIcon from '@material-ui/icons/Close';
-import Slide from '@material-ui/core/Slide';
-import { Delete } from '@material-ui/icons';
+import makeStyles from '@mui/styles/makeStyles';
+import Button from '@mui/material/Button';
+import Dialog from '@mui/material/Dialog';
+import ListItemText from '@mui/material/ListItemText';
+import ListItem from '@mui/material/ListItem';
+import List from '@mui/material/List';
+import Divider from '@mui/material/Divider';
+import AppBar from '@mui/material/AppBar';
+import Toolbar from '@mui/material/Toolbar';
+import IconButton from '@mui/material/IconButton';
+import Typography from '@mui/material/Typography';
+import CloseIcon from '@mui/icons-material/Close';
+import Slide from '@mui/material/Slide';
+import { Delete } from '@mui/icons-material';
 
 const useStyles = makeStyles((theme) => ({
     appBar: {
@@ -30,7 +30,7 @@ const Transition = React.forwardRef(function Transition(props, ref) {
 export default function CartShow({ open, handleClose, cart, setCart }) {
     const clearCart = () => {
         setCart([])
-        console.log("cleared the carrt")
+        console.log("cleared the cart")
     }
     const handleDelete = (itemToDelete) => () => {
         setCart((cart) => cart.filter((cart) => cart.id !== itemToDelete.id));
@@ -38,11 +38,15 @@ export default function CartShow({ open, handleClose, cart, setCart }) {
     const classes = useStyles();
     return (
         <div>
-
             <Dialog fullScreen open={open} onClose={handleClose} TransitionComponent={Transition}>
                 <AppBar className={classes.appBar}>
                     <Toolbar>
-                        <IconButton edge="start" color="inherit" onClick={handleClose} aria-label="close">
+                        <IconButton
+                            edge="start"
+                            color="inherit"
+                            onClick={handleClose}
+                            aria-label="close"
+                            size="large">
                             <CloseIcon />
                         </IconButton>
                         <Typography variant="h6" className={classes.name}>
@@ -61,7 +65,7 @@ export default function CartShow({ open, handleClose, cart, setCart }) {
                         <section key={Math.random()}>
                             <ListItem button>
                                 <ListItemText primary={`Book: ${item.name}`} secondary={`Faculty: ${item.faculty} \n| Price: ${item.discountedPrice}`} />
-                                <IconButton onClick={handleDelete(item)}>
+                                <IconButton onClick={handleDelete(item)} size="large">
                                     <Delete />
                                 </IconButton>
                             </ListItem>
