@@ -1,6 +1,6 @@
 import React from 'react';
-import Skeleton from '@material-ui/lab/Skeleton';
-import { Grid, Box } from '@material-ui/core';
+import Skeleton from '@mui/material/Skeleton';
+import { Grid, Box } from '@mui/material';
 
 export default function Variants() {
     return (
@@ -8,27 +8,27 @@ export default function Variants() {
             <Grid
                 container
                 direction="row"
-                justify="center"
+                justifyContent="center"
                 alignItems="center">
                 <div>
-                    <Skeleton variant="rect" width={210} height={118} />
+                    <Skeleton variant="rectangular" width={210} height={118} />
                     <Skeleton variant="text" animation="pulse" width={210} />
                     <Skeleton variant="text" width={210} />
                     <Skeleton variant="text" width={210} /><br />
                 </div>
                 <div>
-                    <Skeleton variant="rect" width={210} height={118} />
+                    <Skeleton variant="rectangular" width={210} height={118} />
                     <Skeleton variant="text" width={210} />
                     <Skeleton variant="text" width={210} />
                     <Skeleton variant="text" width={210} /><br />
                 </div>
                 <div>
-                    <Skeleton variant="rect" width={210} height={118} />
+                    <Skeleton variant="rectangular" width={210} height={118} />
                     <Skeleton variant="text" width={210} />
                     <Skeleton variant="text" width={210} />
                     <Skeleton variant="text" width={210} /><br />
                 </div>            <div>
-                    <Skeleton variant="rect" width={210} height={118} />
+                    <Skeleton variant="rectangular" width={210} height={118} />
                     <Skeleton variant="text" width={210} />
                     <Skeleton variant="text" width={210} />
                     <Skeleton variant="text" width={210} /><br />
