@@ -18,10 +18,10 @@ import {
     List,
     // Switch,
     //  IconButton
-} from '@material-ui/core';
+} from '@mui/material';
 import Variants from '../Profile/Variant';
-import { makeStyles } from "@material-ui/core/styles";
-import { SchoolRounded, Home, Person } from '@material-ui/icons'
+import makeStyles from '@mui/styles/makeStyles';
+import { SchoolRounded, Home, Person } from '@mui/icons-material'
 // import profileData from './profileData';
 import AppBarred from '../AppBar/AppBarred';
 // import { Link } from 'react-router-dom';
@@ -53,7 +53,7 @@ export default function UserProfile({ isLoggedin, data }) {
                     <Container className="container">
                         <Grid
                             container
-                            justify="center"
+                            justifyContent="center"
                             alignContent="center"
                             alignItems="center"
                             className="row">
@@ -96,6 +96,5 @@ export default function UserProfile({ isLoggedin, data }) {
             </Box>
             <Navigation />
         </Box>
-
-    )
+    );
 }
