@@ -5,14 +5,20 @@ import {
   blue,
   deepPurple,
   deepOrange
-} from "@material-ui/core/colors";
+} from "@mui/material/colors";
 import Home from '../Components/Home/Home';
 import Profile from '../Components/Profile/Profile';
 import Add from '../Components/Add/Add';
 import Login from '../Components/Login/Login';
 // import books from '../Components/data';
 import SignUp from '../Components/Login/SignUp';
-import { ThemeProvider, createMuiTheme, CssBaseline } from '@material-ui/core';
+import {
+  ThemeProvider,
+  StyledEngineProvider,
+  createTheme,
+  CssBaseline,
+  adaptV4Theme,
+} from '@mui/material';
 import MyBooks from '../Components/Profile/myBooks';
 import Edit from '../Components/Profile/Edit';
 import { chain } from "lodash";
@@ -52,7 +58,7 @@ import Categorised from '../Components/Category/Categorised';
 
 function App() {
   const [darkState, setDarkState] = useState(false);
-  const [isLoggedin, setLogin] = useState(false);
+  const [isLoggedin, setLogin] = useState(true);
   // const [SearchField, setSearchField] = useState('');
   const palletType = darkState ? "dark" : "light";
   const mainPrimaryColor = darkState ? orange[500] : blue[500];
@@ -95,9 +101,9 @@ function App() {
     setWish(newWish);
     console.log(newWish);
   }
-  const darkTheme = createMuiTheme({
+  const darkTheme = createTheme(adaptV4Theme({
     palette: {
-      type: palletType,
+      mode: palletType,
       primary: {
         main: mainPrimaryColor,
       },
@@ -105,7 +111,7 @@ function App() {
         main: mainSecondaryColor
       }
     }
-  });
+  }));
   const fuse = new Fuse(filteredResult, {
     keys: ["name", "faculty", "sub_faculty", "description", "year_sem"],
     includeScore: true,
@@ -128,72 +134,74 @@ function App() {
   if (loading) return 'Loading...';
   if (error) return `Error! ${error.message}`;
   return (
-    <ThemeProvider theme={darkTheme}>
-      <CssBaseline />
-      <Router>
-        <Switch>
-          <Route exact path="/">
-            <Home
-              isLoggedin={isLoggedin} 
-              setLogin={setLogin}
-              datas={sBook}
-              Cart={Cart}
-              unique={unique}
-              Wish={Wish}
-              addToCart={addToCart}
-              addToWish={addToWish}
-              setCart={setCart}
-              setWish={setWish}
-              category={category}
-              setCategory={setCategory}
-              setSearchField={onChangeSearch}
-              filteredResults={filteredResult}
-            // error={error}
-            // loading={loading}
-            />
-          </Route>
-          <Route path="/profile">
-            <Profile darkState={darkState} handleThemeChange={handleThemeChange} isLoggedin={isLoggedin} setLogin={setLogin} />
-          </Route>
-          <Route path="/add">
-            <Add isLoggedin={isLoggedin} setLogin={setLogin} />
-          </Route>
-          <Route path="/login">
-            <Login
-              userInfo={userInfo}
-              setUser={setUser}
-              isLoggedin={isLoggedin}
-              setLogin={setLogin}
-            />
-          </Route>
-          <Route path="/signup">
-            <SignUp isLoggedin={isLoggedin} setLogin={setLogin} />
-          </Route>
-          <Route path="/myBooks">
-            <MyBooks />
-          </Route>
-          <Route path="/edit" >
-            <Edit />
-          </Route>
-          <Route path="/user/:id" >
-            <UserProfile data={datas} />
-          </Route>
-          <Route path="/categories" >
-            <Category isLoggedin={isLoggedin} />
-          </Route>
-          <Route path="/books/:id" >
-            <Categorised isLoggedin={isLoggedin} data={data} addToCart={addToCart}
-              addToWish={addToWish}
-              Cart={Cart}
-              Wish={Wish}
-              setCart={setCart}
-              setWish={setWish} />
-          </Route>
-        </Switch>
-        <Navigation />
-      </Router>
+    <StyledEngineProvider injectFirst>
+      <ThemeProvider theme={darkTheme}>
+        <CssBaseline />
+        <Router>
+          <Switch>
+            <Route exact path="/">
+              <Home
+                isLoggedin={isLoggedin} 
+                setLogin={setLogin}
+                datas={sBook}
+                Cart={Cart}
+                unique={unique}
+                Wish={Wish}
+                addToCart={addToCart}
+                addToWish={addToWish}
+                setCart={setCart}
+                setWish={setWish}
+                category={category}
+                setCategory={setCategory}
+                setSearchField={onChangeSearch}
+                filteredResults={filteredResult}
+              // error={error}
+              // loading={loading}
+              />
+            </Route>
+            <Route path="/profile">
+              <Profile darkState={darkState} handleThemeChange={handleThemeChange} isLoggedin={isLoggedin} setLogin={setLogin} />
+            </Route>
+            <Route path="/add">
+              <Add isLoggedin={isLoggedin} setLogin={setLogin} />
+            </Route>
+            <Route path="/login">
+              <Login
+                userInfo={userInfo}
+                setUser={setUser}
+                isLoggedin={isLoggedin}
+                setLogin={setLogin}
+              />
+            </Route>
+            <Route path="/signup">
+              <SignUp isLoggedin={isLoggedin} setLogin={setLogin} />
+            </Route>
+            <Route path="/myBooks">
+              <MyBooks />
+            </Route>
+            <Route path="/edit" >
+              <Edit />
+            </Route>
+            <Route path="/user/:id" >
+              <UserProfile data={datas} />
+            </Route>
+            <Route path="/categories" >
+              <Category isLoggedin={isLoggedin} />
+            </Route>
+            <Route path="/books/:id" >
+              <Categorised isLoggedin={isLoggedin} data={data} addToCart={addToCart}
+                addToWish={addToWish}
+                Cart={Cart}
+                Wish={Wish}
+                setCart={setCart}
+                setWish={setWish} />
+            </Route>
+          </Switch>
+          <Navigation />
+        </Router>
 
-    </ThemeProvider>
+      </ThemeProvider>
+    </StyledEngineProvider>
   );
 }
 export default App;
